Semi Completion
Added a canvas in html. Added text style for the body and added a background color to the canvas in CSS. In JS, i gave the rows, columns, and tiles a size amount. drew the ship and aliens with corresponding coordinates on the canvas. gave the aliens movement velocity as well as the ship on certain keydown events.